opentofu/modules/data_explorer/src/pyproject.toml
[project]
name = "streamlit-app"
version = "1.0.0"
description = "Hello world app"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "duckdb>=1.3.0",
    "google-cloud-storage>=2.19.0",
    "plotly>=6.1.2",
    "streamlit>=1.41.1",
    "streamlit-authenticator>=0.4.1",
]

[tool.uv]
dev-dependencies = [
    "watchdog>=6.0.0",
    "mypy>=1.14.1",
    "ruff>=0.8.4",
    "types-pyyaml>=6.0.12.20241230",
]
